+++ src/schemas/definitions/UnknownSchema.ts
@@ -1,7 +1,0 @@
-import { Schema } from '../Schema.js';
-
-export class UnknownSchema extends Schema {
-    protected _normalize(input: unknown): unknown {
-        return input;
-    }
-}