+++ src/schemas/definitions/index.ts
@@ -1,4 +1,3 @@
-export * from './AnySchema.js';
 export * from './ArraySchema.js';
 export * from './BooleanSchema.js';
 export * from './LiteralSchema.js';
@@ -10,4 +9,3 @@ export * from './StringSchema.js';
 export * from './TupleSchema.js';
 export * from './UndefinedSchema.js';
 export * from './UnionSchema.js';
-export * from './UnknownSchema.js';